feat(web): ✨🚧 add column sorting
